Moved App.js and theme.js into react folder
