+++ src/index.js
@@ -13,7 +13,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App';
+import App from './react/App';
 import store from './redux/store';
 import './index.css';
